part1/requirements.txt
cfn-flip==1.2.1
PyYAML==5.1.2